+++ Kint-Sources-master/OpenNos.Data/ItemInstanceDTO.cs
@@ -119,7 +119,7 @@ namespace OpenNos.Data
 
         public long CharacterId { get; set; }
 
-        public short Design { get; set; }
+        public ushort Design { get; set; }
 
         public int DurabilityPoint { get; set; }
 
@@ -133,7 +133,7 @@ namespace OpenNos.Data
 
         
 
-        public byte Upgrade { get; set; }
+        public ushort Upgrade { get; set; }
 
         #endregion
     }
