+++ Kint-Sources-master/OpenNos.GameObject/Character.cs
@@ -1603,9 +1603,9 @@ namespace OpenNos.GameObject
         public string GenerateEqRareUpgradeForPacket()
         {
             sbyte weaponRare = 0;
-            byte weaponUpgrade = 0;
+            ushort weaponUpgrade = 0;
             sbyte armorRare = 0;
-            byte armorUpgrade = 0;
+            ushort armorUpgrade = 0;
             if (Inventory != null)
             {
                 for (short i = 0; i < 15; i++)
@@ -3245,7 +3245,7 @@ namespace OpenNos.GameObject
             return new[] { value1, value2 };
         }
 
-        public void GiftAdd(short itemVNum, byte amount, byte rare = 0, byte upgrade = 0, short design = 0, bool forceRandom = false, byte minRare = 0)
+        public void GiftAdd(short itemVNum, ushort amount, byte rare = 0, byte upgrade = 0, ushort design = 0, bool forceRandom = false, byte minRare = 0)
         {
             //TODO add the rare support
             if (Inventory != null)
@@ -3926,7 +3926,7 @@ namespace OpenNos.GameObject
             }
         }
 
-        public void SendGift(long id, short vnum, byte amount, sbyte rare, byte upgrade, bool isNosmall)
+        public void SendGift(long id, short vnum, ushort amount, sbyte rare, ushort upgrade, bool isNosmall)
         {
             Item it = ServerManager.GetItem(vnum);
