+++ Kint-Sources-master/OpenNos.GameObject/Packets/ClientPackets/MviPacket.cs
@@ -17,7 +17,7 @@ namespace OpenNos.GameObject
         public short Slot { get; set; }
 
         [PacketIndex(2)]
-        public byte Amount { get; set; }
+        public ushort Amount { get; set; }
 
         [PacketIndex(3)]
         public byte DestinationSlot { get; set; }
